requirements.txt
pandas
tqdm
biopython